+++ Alphamated/assets/.gitignore
@@ -1,0 +1,1 @@
+email_list.txt
