+++ Alphamated/public/contact.html
@@ -97,7 +97,7 @@
                 <div class="mdl-layout-spacer"></div>
 
                 <div class="mdl-card mdl-shadow--2dp my-subscribe-card-size">
-                    <form name="subForm" id="subscriptionForm" method="POST" action="http://alphamated.com/addEmail" autocomplete="off">
+                    <form name="subForm" id="subscriptionForm" method="POST" action="/api/emailList/add" autocomplete="off">
                         <!--todo add code so you dont hard code address above. let env handle-->
                         <div class="mdl-card__title mdl-color--primary">
                             <h2 class="mdl-card__title-text mdl-color-text--white">Subscribe:</h2>
